feat: 화면 가득 채우도록 변경
